--- packages/app/cores/designsystem/lib/font_family.dart
@@ -1,1 +1,0 @@
-export 'src/font_family.dart';
