+++ packages/app/cores/designsystem/lib/providers.dart
@@ -1,0 +1,1 @@
+export 'src/providers/font_family.dart';
